requirements.txt
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Cython==0.29.28
et-xmlfile==1.1.0
fake-useragent==0.1.11
h11==0.13.0
idna==3.3
lxml==4.8.0
numpy==1.22.3
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
PySocks==1.7.1
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tesseract==0.1.3
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
html5lib